STORM-2084: Refactor localization to combine files together
