+++ storm-core/test/clj/integration/org/apache/storm/testing4j_test.clj
@@ -65,13 +65,13 @@
   (reify TestJob
        (^void run [this ^ILocalCluster cluster]
          (let [topology (Thrift/buildTopology
-                         {"1" (Thrift/prepareSpoutDetails (TestWordSpout. true) (Integer. 3))}
+                         {"spout" (Thrift/prepareSpoutDetails (TestWordSpout. true) (Integer. 3))}
                          {"2" (Thrift/prepareBoltDetails
-                                {(GlobalStreamId. "1" Utils/DEFAULT_STREAM_ID)
+                                {(GlobalStreamId. "spout" Utils/DEFAULT_STREAM_ID)
                                  (Thrift/prepareFieldsGrouping ["word"])}
                                 (TestWordCounter.) (Integer. 4))
                           "3" (Thrift/prepareBoltDetails
-                                {(GlobalStreamId. "1" Utils/DEFAULT_STREAM_ID)
+                                {(GlobalStreamId. "spout" Utils/DEFAULT_STREAM_ID)
                                  (Thrift/prepareGlobalGrouping)}
                                 (TestGlobalCount.))
                           "4" (Thrift/prepareBoltDetails
@@ -79,7 +79,7 @@
                                  (Thrift/prepareGlobalGrouping)}
                                 (TestAggregatesCounter.))})
                mocked-sources (doto (MockedSources.)
-                                (.addMockData "1" (into-array Values [(Values. (into-array ["nathan"]))
+                                (.addMockData "spout" (into-array Values [(Values. (into-array ["nathan"]))
                                                                       (Values. (into-array ["bob"]))
                                                                       (Values. (into-array ["joey"]))
                                                                       (Values. (into-array ["nathan"]))])
@@ -93,7 +93,7 @@
                                                  topology
                                                  complete-topology-param)]
            (is (Testing/multiseteq [["nathan"] ["bob"] ["joey"] ["nathan"]]
-                           (Testing/readTuples results "1")))
+                           (Testing/readTuples results "spout")))
            (is (Testing/multiseteq [["nathan" (int 1)] ["nathan" (int 2)] ["bob" (int 1)] ["joey" (int 1)]]
                            (Testing/readTuples results "2")))
            (is (= [[1] [2] [3] [4]]
@@ -102,16 +102,34 @@
                   (Testing/readTuples results "4")))
            ))))
 
-(deftest test-complete-topology
-  (doseq [zmq-on? [true false]
-          :let [daemon-conf (doto (Config.)
-                              (.put STORM-LOCAL-MODE-ZMQ zmq-on?))
-                mk-cluster-param (doto (MkClusterParam.)
-                                   (.setSupervisors (int 4))
-                                   (.setDaemonConf daemon-conf))]]
+(deftest test-complete-topology-netty-simulated
+  (let [daemon-conf (doto (Config.)
+                            (.put STORM-LOCAL-MODE-ZMQ true))
+        mk-cluster-param (doto (MkClusterParam.)
+                                 (.setSupervisors (int 4))
+                                 (.setDaemonConf daemon-conf))]
     (Testing/withSimulatedTimeLocalCluster
-      mk-cluster-param complete-topology-testjob )
+      mk-cluster-param complete-topology-testjob)))
+
+(deftest test-complete-topology-netty
+  (let [daemon-conf (doto (Config.)
+                            (.put STORM-LOCAL-MODE-ZMQ true))
+        mk-cluster-param (doto (MkClusterParam.)
+                                 (.setSupervisors (int 4))
+                                 (.setDaemonConf daemon-conf))]
     (Testing/withLocalCluster
+      mk-cluster-param complete-topology-testjob)))
+
+(deftest test-complete-topology-local
+  (let [mk-cluster-param (doto (MkClusterParam.)
+                                 (.setSupervisors (int 4)))]
+    (Testing/withLocalCluster
+      mk-cluster-param complete-topology-testjob)))
+
+(deftest test-complete-topology-local-simulated
+  (let [mk-cluster-param (doto (MkClusterParam.)
+                                 (.setSupervisors (int 4)))]
+    (Testing/withSimulatedTimeLocalCluster
       mk-cluster-param complete-topology-testjob)))
 
 (deftest test-with-tracked-cluster
