+++ storm-server/src/main/java/org/apache/storm/localizer/LocalDownloadedResource.java
@@ -19,6 +19,7 @@ package org.apache.storm.localizer;
 
 import java.util.HashSet;
 import java.util.Set;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
@@ -30,39 +31,6 @@ import org.slf4j.LoggerFactory;
 
 public class LocalDownloadedResource {
     private static final Logger LOG = LoggerFactory.getLogger(LocalDownloadedResource.class);
-    private static class NoCancelFuture<T> implements Future<T> {
-        private final Future<T> _wrapped;
-        
-        public NoCancelFuture(Future<T> wrapped) {
-            _wrapped = wrapped;
-        }
-        
-        @Override
-        public boolean cancel(boolean mayInterruptIfRunning) {
-            //cancel not currently supported
-            return false;
-        }
-
-        @Override
-        public boolean isCancelled() {
-            return false;
-        }
-
-        @Override
-        public boolean isDone() {
-            return _wrapped.isDone();
-        }
-
-        @Override
-        public T get() throws InterruptedException, ExecutionException {
-            return _wrapped.get();
-        }
-
-        @Override
-        public T get(long timeout, TimeUnit unit) throws InterruptedException, ExecutionException, TimeoutException {
-            return _wrapped.get(timeout, unit);
-        }
-    }
     private static class PortNAssignment {
         private final int _port;
         private final LocalAssignment _assignment;
@@ -91,13 +59,13 @@ public class LocalDownloadedResource {
             return "{"+ _port + " " + _assignment +"}";
         }
     }
-    private final Future<Void> _pending;
+    private final CompletableFuture<Void> _pending;
     private final Set<PortNAssignment> _references;
     private boolean _isDone;
     
     
-    public LocalDownloadedResource(Future<Void> pending) {
-        _pending = new NoCancelFuture<>(pending);
+    public LocalDownloadedResource(CompletableFuture<Void> pending) {
+        _pending = pending;
         _references = new HashSet<>();
         _isDone = false;
     }
@@ -108,7 +76,7 @@ public class LocalDownloadedResource {
      * @param la the assignment this is for
      * @return a future that can be used to track it being downloaded.
      */
-    public synchronized Future<Void> reserve(int port, LocalAssignment la) {
+    public synchronized CompletableFuture<Void> reserve(int port, LocalAssignment la) {
         PortNAssignment pna = new PortNAssignment(port, la);
         if (!_references.add(pna)) {
             LOG.warn("Resources {} already reserved {} for this topology", pna, _references);
